requirements.txt
streamlit
sentry-sdk
openai
promptic
pydantic
pypdf
tenacity
Mastodon.py>=1.8.1
python-dotenv